+++ tests/DatabaseTests/MeritTests/EpochsTests/EDBTest.nim
@@ -40,7 +40,7 @@ import tables
 import random
 
 #Seed random.
-randomize(getTime())
+randomize(int64(getTime()))
 var
     #Database.
     db: DB = newTestDatabase()
