+++ tests/NetworkTests/SerializeTests/Consensus/SerializeSignedVerificationTest.nim
@@ -23,7 +23,7 @@ import ../../../DatabaseTests/ConsensusTests/CompareConsensus
 import random
 
 #Seed random.
-randomize(getTime())
+randomize(int64(getTime()))
 
 var
     #Hash.
