+++ tests/NetworkTests/SerializeTests/Merit/SerializeBlockTest.nim
@@ -30,7 +30,7 @@ import ../../../DatabaseTests/MeritTests/CompareMerit
 import random
 
 #Seed random.
-randomize(getTime())
+randomize(int64(getTime()))
 
 var
     #Last Block's Hash.
@@ -69,8 +69,8 @@ for s in 0 .. 255:
         last,
         newMinerWallet().sign(rand(high(int32)).toBinary()),
         minersHash,
-        rand(high(int32)),
-        rand(high(int32))
+        uint32(rand(high(int32))),
+        uint32(rand(high(int32)))
     )
 
     #Randomize the records.
