+++ tests/NetworkTests/SerializeTests/Merit/SerializeMinersTest.nim
@@ -20,7 +20,7 @@ import ../../../DatabaseTests/MeritTests/CompareMerit
 import random
 
 #Seed Random via the time.
-randomize(getTime())
+randomize(int64(getTime()))
 
 var
     #Miner Seq.
